カロリー評価の実装

+++ project/bodquest_2023/lib/domain/usecase/training/get_calories_consumed_usecase.dart
@@ -1,5 +1,6 @@
 import '../../entity/training.dart';
 
 abstract interface class IGetCaloriesConsumedUsecase {
-  int get(Training target);
+  /// [target]に対応するMetabolic Equivalent of Task(MET)を返します.
+  double get(Training target);
 }
